% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{add_csv_metadata_header}
\alias{add_csv_metadata_header}
\title{Helper function to add metadata header to CSV file}
\usage{
add_csv_metadata_header(file_path, metadata)
}
\description{
Helper function to add metadata header to CSV file
}
\keyword{internal}
