% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtf_parser.R
\name{add_footnote_pattern}
\alias{add_footnote_pattern}
\title{Add footnote indicator pattern}
\usage{
add_footnote_pattern(pattern)
}
\arguments{
\item{pattern}{Regex pattern for footnote detection}
}
\description{
Add footnote indicator pattern
}
