% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtf_parser.R
\name{add_population_pattern}
\alias{add_population_pattern}
\title{Add population pattern}
\usage{
add_population_pattern(pattern)
}
\arguments{
\item{pattern}{Regex pattern for population detection}
}
\description{
Add population pattern
}
