% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pattern_management.R
\name{add_stat_pattern}
\alias{add_stat_pattern}
\title{Add statistic pattern}
\usage{
add_stat_pattern(pattern, stat_names)
}
\arguments{
\item{pattern}{Pattern string}

\item{stat_names}{Vector of statistic names}
}
\description{
Add statistic pattern
}
