% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtf_parser.R
\name{add_table_header_pattern}
\alias{add_table_header_pattern}
\title{Add table header indicator pattern}
\usage{
add_table_header_pattern(pattern)
}
\arguments{
\item{pattern}{Regex pattern for table header detection}
}
\description{
Add table header indicator pattern
}
