% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtf_parser.R
\name{add_title_pattern}
\alias{add_title_pattern}
\title{Add title pattern}
\usage{
add_title_pattern(pattern)
}
\arguments{
\item{pattern}{Regex pattern for title detection}
}
\description{
Add title pattern
}
