% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtf_parser.R
\name{build_column_hierarchy_state}
\alias{build_column_hierarchy_state}
\title{Build column hierarchy using cell positions}
\usage{
build_column_hierarchy_state(header_rows)
}
\description{
Build column hierarchy using cell positions
}
\keyword{internal}
