% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtf_cell_utils.R
\name{calculate_cell_coverage}
\alias{calculate_cell_coverage}
\title{Calculate cell span coverage}
\usage{
calculate_cell_coverage(spanning_cell, covered_cells)
}
\arguments{
\item{spanning_cell}{Cell object that might be spanning}

\item{covered_cells}{List of cell objects that might be covered}
}
\value{
Numeric value between 0 and 1 indicating coverage
}
\description{
For a given cell, calculate what percentage of other cells it covers
Used to identify spanning headers
}
