% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_parser.R
\name{clean_group_text}
\alias{clean_group_text}
\title{Clean group text by removing RTF artifacts}
\usage{
clean_group_text(text)
}
\description{
Clean group text by removing RTF artifacts
}
\keyword{internal}
