% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtf_parser.R
\name{clean_treatment_name}
\alias{clean_treatment_name}
\title{Clean treatment names by extracting just the treatment part}
\usage{
clean_treatment_name(treatment_text)
}
\description{
Clean treatment names by extracting just the treatment part
}
\keyword{internal}
