% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pattern_management.R
\name{configure_table_patterns}
\alias{configure_table_patterns}
\title{Configure table patterns}
\usage{
configure_table_patterns(
  title_patterns = NULL,
  population_patterns = NULL,
  stat_patterns = NULL,
  bign_patterns = NULL,
  footnote_patterns = NULL,
  table_header_patterns = NULL
)
}
\arguments{
\item{title_patterns}{Vector of title patterns}

\item{population_patterns}{Vector of population patterns}

\item{stat_patterns}{List of statistic patterns}

\item{bign_patterns}{List of BIGN patterns}

\item{footnote_patterns}{Vector of footnote patterns}

\item{table_header_patterns}{Vector of table header patterns}
}
\description{
Configure table patterns
}
