% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudo_pattern.R
\name{create_default_labels}
\alias{create_default_labels}
\title{Create default labels from stat names}
\usage{
create_default_labels(stat_names)
}
\description{
Create default labels from stat names
}
\keyword{internal}
