% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtf_to_ard.R
\name{create_empty_ard}
\alias{create_empty_ard}
\title{Create empty ARD data frame}
\usage{
create_empty_ard()
}
\value{
Empty data frame in ARD format
}
\description{
Create an empty ARD data frame with all required columns
}
\keyword{internal}
