% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtf_parser.R
\name{determine_group2_type_state}
\alias{determine_group2_type_state}
\title{Determine group2 type for state-based parsing}
\usage{
determine_group2_type_state(level2_values, config)
}
\description{
Determine group2 type for state-based parsing
}
\keyword{internal}
