% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtf_parser.R
\name{determine_group3_type_state}
\alias{determine_group3_type_state}
\title{Determine group3 type for state-based parsing}
\usage{
determine_group3_type_state(level3_values, config)
}
\description{
Determine group3 type for state-based parsing
}
\keyword{internal}
