% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtf_parser.R
\name{determine_table_end}
\alias{determine_table_end}
\title{Determine where table data ends}
\usage{
determine_table_end(rows, table_start)
}
\description{
Determine where table data ends
}
\keyword{internal}
