% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtf_parser.R
\name{determine_table_end_with_borders}
\alias{determine_table_end_with_borders}
\title{Determine table end using border transitions and header blocks}
\usage{
determine_table_end_with_borders(rows, table_start, header_blocks)
}
\description{
Determine table end using border transitions and header blocks
}
\keyword{internal}
