% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtf_parser.R
\name{determine_table_start_with_borders}
\alias{determine_table_start_with_borders}
\title{Determine table start using border information}
\usage{
determine_table_start_with_borders(rows, header_end)
}
\description{
Determine table start using border information
}
\keyword{internal}
