% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudo_pattern.R
\name{.init_bign_pseudo_patterns}
\alias{.init_bign_pseudo_patterns}
\title{Initialize BIGN pseudo-patterns}
\usage{
.init_bign_pseudo_patterns()
}
\description{
Initialize BIGN pseudo-patterns
}
\keyword{internal}
