% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llm_query.R
\name{execute_ard_sql}
\alias{execute_ard_sql}
\title{Execute SQL query on ARD data}
\usage{
execute_ard_sql(sql, ard_data_list)
}
\arguments{
\item{sql}{SQL query string}

\item{ard_data_list}{Named list of ARD data frames}
}
\value{
Query results as a data frame
}
\description{
Converts ARD data to a temporary SQLite database and executes the query.
}
