% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_parser.R
\name{extract_cell_indentation}
\alias{extract_cell_indentation}
\title{Extract RTF indentation level from cell}
\usage{
extract_cell_indentation(cell)
}
\description{
Extract RTF indentation level from cell
}
\keyword{internal}
