% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtf_parser_working.R
\name{extract_cells_simple}
\alias{extract_cells_simple}
\title{Simple cell extraction for working parser}
\usage{
extract_cells_simple(row_content)
}
\arguments{
\item{row_content}{RTF row content}
}
\value{
Character vector of cell contents
}
\description{
Simple cell extraction for working parser
}
\keyword{internal}
