% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_parser.R
\name{extract_footnotes_from_state_data}
\alias{extract_footnotes_from_state_data}
\title{Extract footnotes from state data}
\usage{
extract_footnotes_from_state_data(state_data)
}
\description{
Extract footnotes from state data
}
\keyword{internal}
