% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtf_pipeline.R
\name{extract_group_from_column}
\alias{extract_group_from_column}
\title{Extract group information from column name}
\usage{
extract_group_from_column(column_name)
}
\description{
Extract group information from column name
}
\keyword{internal}
