% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtf_pipeline.R
\name{extract_population_from_sections}
\alias{extract_population_from_sections}
\title{Extract population from table sections}
\usage{
extract_population_from_sections(table_sections)
}
\description{
Extract population from table sections
}
\keyword{internal}
