% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtf_parser.R
\name{extract_state_metadata}
\alias{extract_state_metadata}
\title{Extract metadata from state-based sections}
\usage{
extract_state_metadata(table_sections, rtf_file)
}
\description{
Extract metadata from state-based sections
}
\keyword{internal}
