% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtf_parser.R
\name{extract_table_data}
\alias{extract_table_data}
\title{Extract table data as clean data frame}
\usage{
extract_table_data(table_rows)
}
\description{
Extract table data as clean data frame
}
\keyword{internal}
