% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtf_parser.R
\name{extract_table_sections}
\alias{extract_table_sections}
\title{Extract sections based on identified states}
\usage{
extract_table_sections(rows, states)
}
\description{
Extract sections based on identified states
}
\keyword{internal}
