% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_parser.R
\name{extract_title_from_state_data}
\alias{extract_title_from_state_data}
\title{Extract title from state data}
\usage{
extract_title_from_state_data(state_data)
}
\description{
Extract title from state data
}
\keyword{internal}
