% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_process_ard.R
\name{extract_total_n}
\alias{extract_total_n}
\title{Extract total N values from ARD data}
\usage{
extract_total_n(ard_data)
}
\arguments{
\item{ard_data}{A data frame in ARD format}
}
\value{
A named vector of total N values by group
}
\description{
Extract total N values from ARD data
}
