% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_parser.R
\name{extract_variable_hierarchy}
\alias{extract_variable_hierarchy}
\title{Extract variable hierarchy from table rows}
\usage{
extract_variable_hierarchy(table_rows)
}
\description{
Extract variable hierarchy from table rows
}
\keyword{internal}
