% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtf_header_utils.R
\name{find_containing_cell}
\alias{find_containing_cell}
\title{Find containing cell for a position}
\usage{
find_containing_cell(position, cell_mapping)
}
\arguments{
\item{position}{Numeric position to search for}

\item{cell_mapping}{List of cell mappings to search in}
}
\value{
Cell object if found, NULL otherwise
}
\description{
Find which cell in a mapping contains a given position
}
\keyword{internal}
