% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtf_cell_utils.R
\name{find_overlapping_cells}
\alias{find_overlapping_cells}
\title{Find cells overlapping with a position range}
\usage{
find_overlapping_cells(cells, start_pos, end_pos)
}
\arguments{
\item{cells}{List of cell objects with start_pos and end_pos}

\item{start_pos}{Numeric start position to check}

\item{end_pos}{Numeric end position to check}
}
\value{
Indices of overlapping cells
}
\description{
Given a list of cells and a position range, find which cells overlap
This is used for mapping data cells to header cells
}
