% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtf_header_utils.R
\name{find_parent_cell}
\alias{find_parent_cell}
\title{Find parent cell for a child cell}
\usage{
find_parent_cell(child_cell, parent_map)
}
\arguments{
\item{child_cell}{Child cell object}

\item{parent_map}{List of parent cell mappings}
}
\value{
Index of parent cell or NULL if not found
}
\description{
Determine which parent cell contains or spans over a child cell
}
\keyword{internal}
