% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llm_query.R
\name{generate_table_schema}
\alias{generate_table_schema}
\title{Generate table schema description from ARD data}
\usage{
generate_table_schema(ard_data)
}
\arguments{
\item{ard_data}{ARD data frame}
}
\value{
Character string describing the schema
}
\description{
Creates a schema description that can be used by the LLM to understand
the structure of the available ARD summary data.
}
