% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pattern_management.R, R/pseudo_pattern.R
\name{get_bign_pseudo_patterns}
\alias{get_bign_pseudo_patterns}
\title{Get BIGN pseudo patterns}
\usage{
get_bign_pseudo_patterns()

get_bign_pseudo_patterns()
}
\value{
List of BIGN pseudo patterns
}
\description{
Get BIGN pseudo patterns

Get BIGN patterns using pseudo-pattern system
}
