% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_structured.R
\name{get_column_info}
\alias{get_column_info}
\title{Get column information for metadata}
\usage{
get_column_info(data)
}
\arguments{
\item{data}{Structured ARD data frame}
}
\description{
Get column information for metadata
}
\keyword{internal}
