% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_parser.R
\name{get_group_info}
\alias{get_group_info}
\title{Get group information for column index}
\usage{
get_group_info(column_index, header_groups)
}
\description{
Get group information for column index
}
\keyword{internal}
