% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pattern_management.R
\name{get_percentage_patterns}
\alias{get_percentage_patterns}
\title{Get percentage patterns (alias for stat patterns)}
\usage{
get_percentage_patterns()
}
\value{
List of percentage patterns
}
\description{
Get percentage patterns (alias for stat patterns)
}
