% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pattern_management.R
\name{get_population_patterns}
\alias{get_population_patterns}
\title{Get population patterns}
\usage{
get_population_patterns()
}
\value{
Vector of population patterns
}
\description{
Get population patterns
}
