% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pattern_management.R
\name{get_table_header_patterns}
\alias{get_table_header_patterns}
\title{Get table header patterns}
\usage{
get_table_header_patterns()
}
\value{
Vector of table header patterns
}
\description{
Get table header patterns
}
