% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R, R/rtf_hierarchy_utils.R
\name{\%||\%}
\alias{\%||\%}
\title{Null-coalescing operator}
\usage{
x \%||\% y

x \%||\% y
}
\arguments{
\item{x}{First value}

\item{y}{Second value}
}
\value{
First non-null value
}
\description{
Return first non-null value
}
\keyword{internal}
