% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtf_parser.R
\name{identify_sections_without_headers}
\alias{identify_sections_without_headers}
\title{Identify sections when no clear headers are found}
\usage{
identify_sections_without_headers(rows)
}
\description{
Identify sections when no clear headers are found
}
\keyword{internal}
