% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtf_parser.R
\name{is_structure_break}
\alias{is_structure_break}
\title{Check if there's a significant structure break}
\usage{
is_structure_break(expected, current)
}
\description{
Check if there's a significant structure break
}
\keyword{internal}
