% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llm_query.R
\name{load_ard_json}
\alias{load_ard_json}
\title{Load ARD JSON file as data frame}
\usage{
load_ard_json(json_file)
}
\arguments{
\item{json_file}{Path to ARD JSON file}
}
\value{
ARD data frame
}
\description{
Convenience function to load ARD JSON files for LLM querying
}
