% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_structured.R
\name{load_structured_ard_json}
\alias{load_structured_ard_json}
\title{Load structured ARD from exported JSON}
\usage{
load_structured_ard_json(json_file)
}
\arguments{
\item{json_file}{Path to JSON file created by export_structured_ard_json()}
}
\value{
Structured ARD object
}
\description{
Utility function to load previously exported structured ARD data.
}
