% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtf_header_utils.R
\name{map_to_parent_level}
\alias{map_to_parent_level}
\title{Map child level to parent level}
\usage{
map_to_parent_level(child_map, parent_map)
}
\arguments{
\item{child_map}{List of child level mappings}

\item{parent_map}{List of parent level mappings}
}
\value{
Enhanced child_map with parent references
}
\description{
For multi-level headers, determine which parent cell each child cell belongs to
}
\keyword{internal}
