% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtf_parser_working.R
\name{parse_rtf_table_states_working}
\alias{parse_rtf_table_states_working}
\title{Parse RTF table using working state machine}
\usage{
parse_rtf_table_states_working(rtf_file)
}
\arguments{
\item{rtf_file}{Path to RTF file}
}
\value{
List with sections compatible with existing JSON outputs
}
\description{
Parse RTF table using working state machine
}
