% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtf_pipeline.R
\name{parse_single_statistic}
\alias{parse_single_statistic}
\title{Parse a single statistic value}
\usage{
parse_single_statistic(stat_value)
}
\description{
Parse a single statistic value
}
\keyword{internal}
