% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtf_parser.R
\name{parse_state_stat_value}
\alias{parse_state_stat_value}
\title{Parse individual statistic values}
\usage{
parse_state_stat_value(stat_value, config = list())
}
\description{
Parse individual statistic values
}
\keyword{internal}
