% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtf_to_ard.R
\name{parse_statistic_value}
\alias{parse_statistic_value}
\title{Parse individual statistic value}
\usage{
parse_statistic_value(stat_value)
}
\arguments{
\item{stat_value}{Character string with statistic}
}
\value{
List of parsed statistics with value, name, and label
}
\description{
Attempt to parse a statistic string into component parts
}
\keyword{internal}
