% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_variable_hierarchy}
\alias{parse_variable_hierarchy}
\title{Parse variable hierarchy for nested structures}
\usage{
parse_variable_hierarchy(data)
}
\arguments{
\item{data}{A data frame in ARD format}
}
\value{
A data frame with parsed variable hierarchy
}
\description{
Supports up to 5 levels of variable nesting (variable1, variable2, etc.)
for complex table structures with multiple classification levels.
}
\keyword{internal}
