% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_parser.R
\name{process_table_rows}
\alias{process_table_rows}
\title{Process table rows to create flat ARD structure}
\usage{
process_table_rows(table_rows, header_groups)
}
\description{
Process table rows to create flat ARD structure
}
\keyword{internal}
