% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtf_header_utils.R
\name{rtf-header-utils}
\alias{rtf-header-utils}
\title{RTF Header Structure Analysis Utilities}
\description{
Utilities for analyzing multi-level header structures in RTF tables
and building column group mappings
}
