% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtf_parser_working.R
\name{rtf-parser-working}
\alias{rtf-parser-working}
\title{Working RTF Parser (Restored Functionality)}
\description{
Restore the original working parser functionality that created the existing JSON outputs
}
