% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtf_parser.R
\name{set_bign_patterns}
\alias{set_bign_patterns}
\title{Set BIGN patterns}
\usage{
set_bign_patterns(patterns)
}
\arguments{
\item{patterns}{Vector of regex patterns (from most to least restrictive)}
}
\description{
Updates the global BIGN patterns used for extraction.
}
