% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pattern_management.R
\name{set_stat_patterns}
\alias{set_stat_patterns}
\title{Set statistic patterns}
\usage{
set_stat_patterns(patterns)
}
\arguments{
\item{patterns}{List of statistic patterns}
}
\description{
Set statistic patterns
}
