% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pattern_management.R
\name{add_bign_pattern}
\alias{add_bign_pattern}
\title{Add BIGN pattern}
\usage{
add_bign_pattern(pattern, stat_name = "n")
}
\arguments{
\item{pattern}{Pattern string}

\item{stat_name}{Statistic name (defaults to "n")}
}
\description{
Add BIGN pattern
}
