% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pattern_management.R
\name{add_footnote_pattern}
\alias{add_footnote_pattern}
\title{Add footnote pattern}
\usage{
add_footnote_pattern(pattern)
}
\arguments{
\item{pattern}{Pattern string}
}
\description{
Add footnote pattern
}
