% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pattern_management.R
\name{add_table_header_pattern}
\alias{add_table_header_pattern}
\title{Add table header pattern}
\usage{
add_table_header_pattern(pattern)
}
\arguments{
\item{pattern}{Pattern string}
}
\description{
Add table header pattern
}
