% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtf_parser.R
\name{configure_table_patterns}
\alias{configure_table_patterns}
\title{Configure patterns for a new table type}
\usage{
configure_table_patterns(
  table_type,
  title_patterns = NULL,
  population_patterns = NULL,
  footnote_patterns = NULL,
  header_patterns = NULL
)
}
\arguments{
\item{table_type}{Name of the table type (e.g., "safety", "efficacy", "demographics")}

\item{title_patterns}{Vector of regex patterns for title detection}

\item{population_patterns}{Vector of regex patterns for population detection}

\item{footnote_patterns}{Vector of regex patterns for footnote detection}

\item{header_patterns}{Vector of regex patterns for table header detection}
}
\description{
Easily add patterns for new table types without hardcoding
}
