% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudo_pattern.R
\name{get_bign_pseudo_patterns}
\alias{get_bign_pseudo_patterns}
\title{Get BIGN patterns using pseudo-pattern system}
\usage{
get_bign_pseudo_patterns()
}
\description{
Get BIGN patterns using pseudo-pattern system
}
\keyword{internal}
