% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtf_parser.R
\name{parse_rtf_table_states}
\alias{parse_rtf_table_states}
\title{Parse RTF table with state identification}
\usage{
parse_rtf_table_states(rtf_file)
}
\arguments{
\item{rtf_file}{Path to RTF file}
}
\value{
List with sections: pre_header, header, table, footnotes
}
\description{
Parse RTF table with state identification
}
