% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pattern_management.R
\name{set_bign_patterns}
\alias{set_bign_patterns}
\title{Set BIGN patterns}
\usage{
set_bign_patterns(patterns)
}
\arguments{
\item{patterns}{List of BIGN patterns}
}
\description{
Set BIGN patterns
}
